// sistema-de-gestao/frontend/src/types.ts
export interface Client {
  id: string;
  name: string;
  logoUrl: string;
  active?: boolean;
  Criado?: string;
  Contrato?: string;
  pais?: string;
  contato_principal?: string;
  cnpj?: string;
  telefone?: string;
  tipo_cliente?: 'parceiro' | 'cliente_final';
  partner_id?: string;
  responsavel_interno_id?: string;
  responsavel_externo?: string;
  email_contato?: string;
}